[TM-3426] remove subclass and add paranoid as include soft delete
